+++ app/locales/en/translations.js
@@ -58,6 +58,8 @@ export default {
         version: 'Version',
         preprint_doi: `{{preprintWords.Preprint}} DOI`,
         article_doi: `Peer-reviewed Publication DOI`,
+        preprint_pending_doi: `DOI created after {{preprintWords.preprint}} is made public`,
+        preprint_pending_doi_moderation: `DOI created after moderator approval`,
         preprint_pending_doi_minted: `DOIs are minted by a third party, and may take up to 24 hours to be registered.`,
         original_publication_date: `Original Publication Date`,
         citations: `Citations`,
@@ -120,7 +122,14 @@ export default {
             example: `See an example`
         },
         subjects: {
-            heading: `Browse <small>by {{browse-by}}</small>`
+            heading: {
+                provider: `Browse <small>by provider</small>`,
+                hasHighlightedSubjects: `Browse <small>by featured subjects</small>`,
+                noHighlightedSubjects: `Browse <small>by subjects</small>`,
+            },
+            links: {
+                seeAllSubjects: `See all subjects available`,
+            },
         },
         services: {
             top: {